oracle/requirements.txt
cbor2
flake8
py-solc-x
pytest
oasis-sapphire-py
ollama
httpx
bech32
